+++ spec/chef/recipes/postgresql_spec.rb
@@ -5,7 +5,7 @@ describe 'postgresql' do
 
   before { allow(Gitlab).to receive(:[]).and_call_original }
 
-  context 'when shared_preload_libraries is not configured' do
+  context 'with default settings' do
     it 'correctly sets the shared_preload_libraries default setting' do
       expect(chef_run.node['gitlab']['postgresql']['shared_preload_libraries'])
         .to be_nil
@@ -13,11 +13,22 @@ describe 'postgresql' do
       expect(chef_run).to render_file('/var/opt/gitlab/postgresql/data/postgresql.conf')
         .with_content(/shared_preload_libraries = ''/)
     end
+
+    it 'correctly sets the log_line_prefix default setting' do
+      expect(chef_run.node['gitlab']['postgresql']['log_line_prefix'])
+        .to be_nil
+
+      expect(chef_run).to render_file('/var/opt/gitlab/postgresql/data/postgresql.conf')
+        .with_content(/log_line_prefix = ''/)
+    end
   end
 
-  context 'when shared_preload_libraries is set' do
+  context 'when user settings are set' do
     before do
-      stub_gitlab_rb(postgresql: { shared_preload_libraries: 'pg_stat_statements'})
+      stub_gitlab_rb(postgresql: {
+        shared_preload_libraries: 'pg_stat_statements',
+        log_line_prefix: '%a'
+        })
     end
 
     it 'correctly sets the shared_preload_libraries setting' do
@@ -26,6 +37,14 @@ describe 'postgresql' do
 
       expect(chef_run).to render_file('/var/opt/gitlab/postgresql/data/postgresql.conf')
         .with_content(/shared_preload_libraries = 'pg_stat_statements'/)
+    end
+
+    it 'correctly sets the log_line_prefix setting' do
+      expect(chef_run.node['gitlab']['postgresql']['log_line_prefix'])
+        .to eql('%a')
+
+      expect(chef_run).to render_file('/var/opt/gitlab/postgresql/data/postgresql.conf')
+        .with_content(/log_line_prefix = '%a'/)
     end
   end
 end